+++ packages/loader/src/Texture2DContentRestorer.ts
@@ -1,0 +1,23 @@
+import { AssetPromise, ContentRestorer, request, Texture2D } from "@oasis-engine/core";
+import { RequestConfig } from "@oasis-engine/core/types/asset/request";
+
+/**
+ * @internal
+ */
+export class Texture2DContentRestorer extends ContentRestorer<Texture2D> {
+  constructor(resource: Texture2D, public url: string, public requestConfig: RequestConfig) {
+    super(resource);
+  }
+
+  /**
+   * @override
+   */
+  restoreContent(): AssetPromise<Texture2D> {
+    return request<HTMLImageElement>(this.url, this.requestConfig).then((image) => {
+      const resource = this.resource;
+      resource.setImageSource(image);
+      resource.generateMipmaps();
+      return resource;
+    });
+  }
+}
