+++ packages/loader/src/TextureCubeLoader.ts
@@ -1,16 +1,15 @@
 import {
   AssetPromise,
   AssetType,
-  ContentRestorer,
   Loader,
   LoadItem,
-  request,
   resourceLoader,
   ResourceManager,
   TextureCube,
   TextureCubeFace
 } from "@oasis-engine/core";
 import { RequestConfig } from "@oasis-engine/core/types/asset/request";
+import { TextureCubeContentRestorer } from "./TextureCubeContentRestorer";
 
 @resourceLoader(AssetType.TextureCube, [""])
 class TextureCubeLoader extends Loader<TextureCube> {
@@ -42,32 +41,6 @@ class TextureCubeLoader extends Loader<TextureCube> {
 
           resourceManager.addContentRestorer(new TextureCubeContentRestorer(texture, urls, requestConfig));
           resolve(texture);
-        })
-        .catch((e) => {
-          reject(e);
-        });
-    });
-  }
-}
-
-class TextureCubeContentRestorer extends ContentRestorer<TextureCube> {
-  constructor(resource: TextureCube, public urls: string[], public requestConfig: RequestConfig) {
-    super(resource);
-  }
-
-  /**
-   * @override
-   */
-  restoreContent(): AssetPromise<TextureCube> {
-    return new AssetPromise((resolve, reject) => {
-      Promise.all(this.urls.map((url) => request<HTMLImageElement>(url, this.requestConfig)))
-        .then((images) => {
-          const resource = this.resource;
-          for (let faceIndex = 0; faceIndex < 6; faceIndex++) {
-            resource.setImageSource(TextureCubeFace.PositiveX + faceIndex, images[faceIndex], 0);
-          }
-          resource.generateMipmaps();
-          resolve(resource);
         })
         .catch((e) => {
           reject(e);
